refactor(ui): further flatten `event-flow/` dir
- I thought the `event-flow-details/` dir might make sense as an abstraction, but then I realized:
  1. `id.ts`, for instance, is also used in `event-sources/` and `sensors`, so not exactly abstracted inside then
  2. it's the only real component in `event-flow/`, since the container is just a routing wrapper
