+++ ui/src/app/event-flow/event-flow-container.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {Route, RouteComponentProps, Switch} from 'react-router';
 
-import {EventFlowPage} from './event-flow-details/event-flow-page';
+import {EventFlowPage} from './event-flow-page';
 
 export const EventFlowContainer = (props: RouteComponentProps<any>) => (
     <Switch>
