+++ ui/src/app/event-flow/event-flow-page.tsx
@@ -5,30 +5,30 @@ import {RouteComponentProps} from 'react-router-dom';
 import {Observable} from 'rxjs';
 import {filter, map} from 'rxjs/operators';
 
-import {kubernetes, Workflow} from '../../../models';
-import {EventSource} from '../../../models/event-source';
-import {Sensor} from '../../../models/sensor';
-import {uiUrl} from '../../shared/base';
-import {Button} from '../../shared/components/button';
-import {ErrorNotice} from '../../shared/components/error-notice';
-import {InfoIcon} from '../../shared/components/fa-icons';
-import {GraphPanel} from '../../shared/components/graph/graph-panel';
-import {Node} from '../../shared/components/graph/types';
-import {Links} from '../../shared/components/links';
-import {NamespaceFilter} from '../../shared/components/namespace-filter';
-import {ResourceEditor} from '../../shared/components/resource-editor/resource-editor';
-import {useCollectEvent} from '../../shared/components/use-collect-event';
-import {ZeroState} from '../../shared/components/zero-state';
-import {Context} from '../../shared/context';
-import {Footnote} from '../../shared/footnote';
-import {historyUrl} from '../../shared/history';
-import {ListWatch} from '../../shared/list-watch';
-import {RetryObservable} from '../../shared/retry-observable';
-import {services} from '../../shared/services';
-import {useQueryParams} from '../../shared/use-query-params';
-import {Utils} from '../../shared/utils';
-import {EventsPanel} from '../../workflows/components/events-panel';
-import {FullHeightLogsViewer} from '../../workflows/components/workflow-logs-viewer/full-height-logs-viewer';
+import {kubernetes, Workflow} from '../../models';
+import {EventSource} from '../../models/event-source';
+import {Sensor} from '../../models/sensor';
+import {uiUrl} from '../shared/base';
+import {Button} from '../shared/components/button';
+import {ErrorNotice} from '../shared/components/error-notice';
+import {InfoIcon} from '../shared/components/fa-icons';
+import {GraphPanel} from '../shared/components/graph/graph-panel';
+import {Node} from '../shared/components/graph/types';
+import {Links} from '../shared/components/links';
+import {NamespaceFilter} from '../shared/components/namespace-filter';
+import {ResourceEditor} from '../shared/components/resource-editor/resource-editor';
+import {useCollectEvent} from '../shared/components/use-collect-event';
+import {ZeroState} from '../shared/components/zero-state';
+import {Context} from '../shared/context';
+import {Footnote} from '../shared/footnote';
+import {historyUrl} from '../shared/history';
+import {ListWatch} from '../shared/list-watch';
+import {RetryObservable} from '../shared/retry-observable';
+import {services} from '../shared/services';
+import {useQueryParams} from '../shared/use-query-params';
+import {Utils} from '../shared/utils';
+import {EventsPanel} from '../workflows/components/events-panel';
+import {FullHeightLogsViewer} from '../workflows/components/workflow-logs-viewer/full-height-logs-viewer';
 import {buildGraph} from './build-graph';
 import {genres} from './genres';
 import {ID} from './id';
