+++ ui/src/app/sensors/components/sensor-details/sensor-details.tsx
@@ -4,7 +4,7 @@ import {useContext, useEffect, useState} from 'react';
 import {RouteComponentProps} from 'react-router';
 
 import {Sensor} from '../../../../models';
-import {ID} from '../../../event-flow/components/event-flow-details/id';
+import {ID} from '../../../event-flow/id';
 import {uiUrl} from '../../../shared/base';
 import {ErrorNotice} from '../../../shared/components/error-notice';
 import {Node} from '../../../shared/components/graph/types';
